Add files via upload
This code creates a portal from where you are standing to coordinates provided in the script. You can use this code in Minecraft Java Edition 1.12. It's based on the course "unga programmerare" (young programmers) by Tobias Hallberg. You have to use a server in Minecraft Java Edition 1.12 to use the code.